% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{przyklad}
\alias{przyklad}
\title{Przykladowe dane.}
\format{
Ramka danych ze 100 wierszami i 6 kolumnami:
\itemize{
  \item{\code{id} - identyfikator obserwacji,}
  \item{\code{zm1} - zmienna numeryczna, zawiera braki danych,}
  \item{\code{zm2} - czynnik (\code{factor}), bez braków danych,}
  \item{\code{zm3} - czynnik (\code{factor}), zawiera danych,}
  \item{\code{zm4} - zmienna etykietowana (\code{haven_labelled}), bez braków
        danych,}
  \item{\code{waga} - zmienna numeryczna, która może być wykorzystana jako
        wagi obserwacji.}
}
}
\usage{
przyklad
}
\description{
Całkowicie artefaktualny zbiór, zawierający niezależnie od
siebie wygenerowane zmienne, za to różnych typów. Pozwala pokazać, jak
zachowują się funkcje pakietu, gdy są wywoływane z kolumnami różnych typów.
}
\keyword{datasets}
