% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab.data.frame}
\alias{tab.tbl_svy}
\alias{tab.default}
\alias{print.tab_lbl}
\title{Tabela z rozkladem dla zmiennych etykietowanych}
\usage{
tab(x, ..., procenty = TRUE, d = 1, suma = TRUE, etykietaSuma, etykietaBD)

\method{tab}{data.frame}(
  x,
  ...,
  procenty = TRUE,
  d = 1,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{tbl_svy}(
  x,
  ...,
  procenty = TRUE,
  d = 1,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{default}(
  x,
  ...,
  procenty = TRUE,
  d = 1,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{print}{tab_lbl}(x, ...)
}
\arguments{
\item{x}{obiekt klasy \code{labelled}, wektor, ramka danych lub obiekt
klasy \code{tbl_svy}}

\item{...}{tylko jeśli \code{x} jest \emph{ramką danych} - kolumna, której
rozkład ma zostać wygenerowany}

\item{procenty}{wartość logiczna - czy wyświetlić również rozkład częstości?}

\item{d}{liczba całkowita - liczba miejsc dziesiętnych, z jaką raportowane
będą procenty (argument jest ignowowany, jeśli \code{procenty} równe
\code{FALSE})}

\item{suma}{wartość logiczna - czy wyświetlić wiersz z sumą?}

\item{etykietaSuma}{ciąg znaków - etykieta dla wiersza z sumą (argument jest
ignorowany, jeśli \code{suma} równe \code{FALSE})}

\item{etykietaBD}{ciąg znaków - etykieta, którą w przygotowanym zestawieniu
mają być opisane braki danych (\code{NA}); domyślna wartość oznacza, że
zostaną one opisane jako "NA"; podanie \code{NULL} będzie skutkować
usunięciem kolumn i wierszy opisujących braki danych ze zwracanego
zestawienia}
}
\value{
\code{data.frame} (klasy \code{tab_lbl}) z rozkładami
}
\description{
Funkcja generuje rozkład jednej zmiennej - liczebności i częstości -
uwzględniając etykiety wartości.
}
