% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2.R, R/tab2_metody.R
\name{tab2}
\alias{tab2}
\alias{tab2.data.frame}
\alias{tab2.table}
\alias{tab2.ftable}
\alias{tab2.tbl_svy}
\alias{tab2.survey.design2}
\alias{tab2.svyrep.design}
\alias{tab2.twophase2}
\alias{print.tab_lbl2}
\alias{as_tibble.tab_lbl2}
\alias{as.data.frame.tab_lbl2}
\title{Tabela z rodzina rozkladow warunkowych dwoch zmiennych (etykietowanych)}
\usage{
tab2(
  x,
  zmW,
  zmK,
  sumowanie,
  liczby,
  procenty,
  etykietaSuma,
  etykietaOgolem,
  etykietaBD,
  ...
)

\method{tab2}{data.frame}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...,
  w = NULL
)

\method{tab2}{table}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{tab2}{ftable}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{tab2}{tbl_svy}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{tab2}{survey.design2}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{tab2}{svyrep.design}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{tab2}{twophase2}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "SUMA",
  etykietaOgolem = "OGÓŁEM",
  etykietaBD = NA,
  ...
)

\method{print}{tab_lbl2}(x, dProcenty = 1, dLiczby = 0, decimal.mark = ",", scipen = 100, ...)

\method{as_tibble}{tab_lbl2}(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = NULL,
  usunSuma = FALSE,
  usunOgolem = FALSE,
  wartoscBD = NULL
)

\method{as.data.frame}{tab_lbl2}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  usunSuma = FALSE,
  usunOgolem = FALSE,
  wartoscBD = NULL,
  niePrzeksztalcaj = TRUE
)
}
\arguments{
\item{x}{ramka danych lub obiekt klasy \code{tbl_svy}}

\item{zmW}{kolumna obiektu \code{x}, której wartości mają zostać umieszczone
w wierszach zwróconego rozkładu}

\item{zmK}{kolumna obiektu \code{x}, której wartości mają zostać umieszczone
w kolumnach zwróconego rozkładu}

\item{sumowanie}{w którą stronę, jeśli w ogóle, powinno zostać dokonane
sumowanie lub procentowanie? (wystarczy podać pierwszą literę)}

\item{liczby}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
liczebności?}

\item{procenty}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
częstości?}

\item{etykietaSuma}{ciąg znaków - etykieta dla wiersza lub kolumy z sumą
(argument jest ignorowany, jeśli \code{sumowanie} równe \code{"brak"} lub
\code{NULL})}

\item{etykietaOgolem}{ciąg znaków - etykieta dla wiersza lub kolumny
z rozkladem brzegowym zmiennej, której rodzina rozkładów warunkowowych jest
tworzona (argument jest ignorowany, jeśli \code{sumowanie} równe
\code{"brak"}, \code{"ogółem"} lub \code{NULL}); podanie wartości \code{NA}
lub \code{NULL} oznacza, że odpowiedni wiersz/kolumna nie powinna znaleźć
się w zwróconym zestawieniu}

\item{etykietaBD}{ciąg znaków - etykieta, którą w przygotowanym zestawieniu
mają być opisane braki danych (\code{NA}); domyślna wartość oznacza, że
zostaną one opisane jako "NA"; podanie \code{NULL} będzie skutkować
usunięciem kolumn i wierszy opisujących braki danych ze zwracanego
zestawienia}

\item{...}{tylko jeśli \code{x} jest \emph{ramką danych} - kolumna, której
rozkład ma zostać wygenerowany}

\item{w}{opcjonalnie kolumna obiektu \code{x}, której wartości zawierają
wagi obserwacji, które powinny zostać uwzględnione przy obliczaniu rozkładu}

\item{dProcenty}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami częstości}

\item{dLiczby}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami liczebności (to mogą być liczby niecałkowite,
jeśli przy tworzeniu rozkładu stosowano ważenie)}

\item{decimal.mark}{znak miejsca dziesiętnego - przekazywany do
\code{\link[base]{format}}}

\item{scipen}{liczba całkowita decydująca o skłonności R do zapisywania liczb
w notacji naukowej (czym większa, tym rzadziej R sosuje notację naukową - p.
\code{\link[base]{options}})}

\item{.rows}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as_tibble}}

\item{.name_repair}{przekazywana jako argument \code{names_repair} do
\code{\link[tidyr]{pivot_longer}}}

\item{rownames}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as_tibble}}

\item{usunSuma}{wartość logiczna - czy usunąć ze zwróconej ramki danych
wiersze opisujące sumy rozkładów?}

\item{usunOgolem}{wartość logiczna - czy usunąć ze zwróconej ramki danych
wiersze opisujące rozkład brzegowy zmiennej zależnej?}

\item{wartoscBD}{wektor jedno- lub dwuelementowy: wartości, przy pomocy
których mają być reprezentowane ew. braki danych w zróconej ramce danych;
jeśli zostanie podany tylko jeden element, zostanie użyty zarówno
w odniesieniu do zmiennej prezentowanej w wierszach, jak i zmiennej
prezentowanej w kolumnach; jeśli zostaną podane dwa elementy, pierwszy
zostanie zastosowany do zmiennej prezentowanej w wierszach, a drugi do
zmiennej prezentowanej w kolumnach}

\item{row.names}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as.data.frame}}

\item{optional}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as.data.frame}}

\item{niePrzeksztalcaj}{wartość logiczna - czy funkcja ma tylko usunąć
z obiektu klasę \code{tab_lbl2} i zwrócić go jako \emph{zwykłą} ramkę danych,
\strong{bez} dokonywania przekształcenia z postaci \emph{szerokiej} do
\emph{długiej})? (domyślnie funkcja dokonuje przekształcenia)}
}
\value{
\strong{tab2}
\code{data.frame} (klasy \code{tab_lbl2}) z rozkładami:
\itemize{
  \item{pierwsza kolumna zawiera wartości zmiennej podanej argumentem
        \code{zmW} i ma taką samą nazwę, jak ta zmienna,}
  \item{nazwy kolejnych kolumn zaczynające się od "n_" opisują rozkład
        liczebności,}
  \item{nazwy kolejnych kolumn zaczynające się od "pct_" opisują rozkład
        częstości,}
  \item{nazwy tych kolumn powstały przez połącznie przedrostka "n_" lub
        "pct_" z wartościami zmiennej podanej argumentem \code{zmK} oraz
        ew. etykietą podaną argumentem \code{etkietaSuma}.}
}
Pierwsza zwracana kolumna rozkładu co do zasady jest \emph{czynnikiem}, chyba
że jednocześnie:
\itemize{
  \item{jako argument \code{sumowanie} zostało podane "brak" lub "wiersze"
        (ew. \code{NULL}),}
  \item{nie został podany argument \code{etykietaSumy} (ew. został podany
        jako \code{NA}),}
  \item{kolumna podana argumentem \code{zmW} jest klasy "character",
        "integer", "numeric" lub "logical".}
}
W takim przypadku będzie mieć ona takiego typu, jaki kolumna miała w danych
wejściowych.
\strong{metoda as_tibble}
W zależności od wartości parametru \code{niePrzeksztalcaj}:
\itemize{
  \item{\emph{tibble} z rozkładami przekształconymi do postaci \emph{długiej},}
  \item{obiekt klasy \code{data.frame} (i tylko tej jednej) z rozkładami
        w postacie \emph{szerokiej}.}
}
}
\description{
Funkcja generuje rozkład łączny liczebności i rozkład łączny lub rodzinę
warunkowych rozkładów częstości dwóch zmiennych. Jako pierwszy argument
przyjmuje obiekt \code{data.frame} lub \code{tbl_svy} (obiekt będący
połączeniem \emph{ramki danych} z informacjami o - zwykle złożonym -
schemacie doboru próby, tworzonym przez funkcje pakietu \emph{srvyr}). Radzi
też sobie ze zmiennymi etykietowanymi (konwertując je na \emph{czynniki}).

Metoda \code{as_tibble} pozwala przekształcić zwracane zestawienie
w ramkę danych w postaci \emph{długiej}, przydatną np. do rysowania wykresów
przy pomocy \emph{ggplot2} (czyli działa analogicznie,jak metoda
\code{as.data.frame} dla obiektów zwracanych przez funkcję
\code{\link[base]{table}} z pakietu \emph{base}). Metoda \code{as.data.frame}
również jest dostępna, ale domyślnie nie przekształca tabeli (usuwa jedynie
klase \code{tab_n}), gdyż inaczej przygotowanych tabel z rozkładami nie
dawałyby się bezproblemowo używać w połączeniu z funkcją
\code{\link[knitr]{kable}} z pakietu \emph{kable} (która to funkcja wywołuje
wywołuje metodę \code{as.data.frame} na przekazywanym jej obiekcie).
}
\examples{
mtcars$cyl = haven::labelled(mtcars$cyl,
                             c("cztery" = 4, "sześć" = 6, "osiem" = 8),
                             "liczba cylindrów")
tab2(mtcars, carb, cyl, "k")
tab2(mtcars, carb, cyl, "w")
tab2(mtcars, carb, cyl, "o")
tab2(mtcars, carb, cyl)
tab2(mtcars, carb, cyl, "k", liczby = FALSE)
tab2(mtcars, carb, cyl, "k", procenty = FALSE)
tab2(mtcars, carb, cyl, "k", etykietaSuma = "Suma")
mtcars$carb[1:2] = NA
tab2(mtcars, carb, cyl, "k")
tab2(mtcars, carb, cyl, "k", etykietaBD = "brak danych")

# choć hp jako wagi obserwacji są bez sensu
tab2(mtcars, carb, cyl, "k", w = hp)

as.data.frame(tab2(mtcars, cyl, carb, "k"))
# aby uzyskać obiekt nadający się do wykorzystania
# jako źródło danych do wykresu tworzonego przy pomocy pakietu ggplot2:
dplyr::as_tibble(tab2(mtcars, cyl, carb, "k"), usunSuma = TRUE)
# lub
dplyr::as_tibble(tab2(mtcars, cyl, carb, "k"), usunSuma = TRUE, usunOgolem = TRUE)
}
