% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_w.R
\name{tab_w}
\alias{tab_w}
\title{Tabela z rozkladem wielokrotnych odpowiedzi}
\usage{
tab_w(
  x,
  wybrane,
  wyklucz,
  kierunek = c("kolumny", "wiersze"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaOgolem = "OGÓŁEM",
  etykietyPozycji = TRUE,
  prefiksPozycji = "",
  wykluczBD = TRUE
)
}
\arguments{
\item{x}{ramka danych lub obiekt klasy \code{tbl_svy}}

\item{wybrane}{wektor wartości, które wskazują na to, że pozycja została
wybrana (jego elementami mogą być ciągi znaków, nazwy podawane jako wyrażenia
języka oraz selektory z pakietu \emph{dplyr}, np. \code{\link{starts_with}})}

\item{wyklucz}{wektor wartości, które \strong{nie} powinny być wliczane do
podstawy procentowania (jego elementami mogą być ciągi znaków, nazwy podawane
jako wyrażenia języka oraz selektory z pakietu \emph{dplyr}, np.
\code{\link{starts_with}})}

\item{kierunek}{w którą stronę mają układać się w zwracanym zestawieniu
pozycje kafeterii?}

\item{liczby}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
liczebności?}

\item{procenty}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
częstości?}

\item{etykietaOgolem}{ciąg znaków - etykieta dla wiersza lub kolumy z łączną
liczbą obserwacji (podstawą procentowania)}

\item{etykietyPozycji}{wartość logiczna - czy do opisania pozycji w kafterii
w zwracanym zestawieniu mają zostać wykorzystane etykiety (o ile są
dostępne)? jeśli \code{FALSE}, użyte zostaną nazwy odpowiednich zmiennych}

\item{prefiksPozycji}{ciąg znaków - część etykiet zmiennych (pozycji), które
opisują treść główki pytania - jeśli podany, zostanie usunięty z etykiet
pozycji i ustawiony jako etykieta całej tabeli}

\item{wykluczBD}{wartość logiczna - czy z podstawy procentowania wykluczać
również braki danych?}
}
\value{
\strong{tab_w}
\code{data.frame} (klasy \code{tab_lbl_n}) z rozkładami:

\strong{metoda as_tibble}
W zależności od wartości parametru \code{niePrzeksztalcaj}:
\itemize{
  \item{\emph{tibble} z rozkładami przekształconymi do postaci \emph{długiej},}
  \item{obiekt klasy \code{data.frame} (i tylko tej jednej) z rozkładami
        w postacie \emph{szerokiej}.}
}
}
\description{
Na podstawie obiektu zwróconego przez funkcję \code{\link{tab_n}} funkcja
generuje tabelę zawierającą rozkład odpowiedzi na pytanie, w którym można
było wybrać więcej niż jedną pozycję kafeterii (tzw. wielokrotne odpowiedzi).

Metoda \code{as_tibble} pozwala przekształcić zwracane zestawienie
w ramkę danych w postaci \emph{długiej}, przydatną np. do rysowania wykresów
przy pomocy \emph{ggplot2} (czyli działa analogicznie,jak metoda
\code{as.data.frame} dla obiektów zwracanych przez funkcję
\code{\link[base]{table}} z pakietu \emph{base}). Metoda \code{as.data.frame}
również jest dostępna, ale domyślnie nie przekształca tabeli (usuwa jedynie
klasę \code{tab_n}), gdyż inaczej przygotowanych tabel z rozkładami nie
dawałoby się bezproblemowo używać w połączeniu z funkcją
\code{\link[knitr]{kable}} z pakietu \emph{kable} (która to funkcja wywołuje
metodę \code{as.data.frame} na przekazywanym jej obiekcie).
}
\examples{
polprodukt = tab_n(przykladWO, starts_with("P1_"))
polprodukt

tab_w(polprodukt, "T")
tab_w(polprodukt, "T", wykluczBD = FALSE)
tab_w(polprodukt, "T", prefiksPozycji = "Jakie kwiaty lubisz:")
tab_w(polprodukt, "T", kierunek = "w", prefiksPozycji = "Jakie kwiaty lubisz:")
tab_w(polprodukt, "T", procenty = FALSE)
tab_w(polprodukt, "T", etykietaOgolem = "Ogółem")

# nieco bardziej skomplikowana sytuacja
polprodukt2 = tab_n(przykladWO, starts_with("P2_"))
polprodukt2
tab_w(polprodukt2, c("T", 1, "Tak"))

# nie ma to tutaj merytorycznego sensu, i prowadzi do ostrzeżenia ale jako
# przykład, że można podać wartości, które mają zostać wykluczone z podstawy
#procentowania
tab_w(polprodukt2, c("T", 1, "Tak"), c(starts_with("n"), 0))

# aby uzyskać obiekt nadający się do wykorzystania
# jako źródło danych do wykresu tworzonego przy pomocy pakietu ggplot2:
dplyr::as_tibble(tab_w(polprodukt2, c("T", 1, "Tak")))
}
