% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zastosuj_codebook.R
\name{zastosuj_codebook}
\alias{zastosuj_codebook}
\title{Utworzenie ramki danych ze zmiennymi etykietowanymi z pliku CSV}
\usage{
zastosuj_codebook(
  dane,
  codebook,
  encoding = "windows-1250",
  etykietyBrakowDanych = ""
)
}
\arguments{
\item{dane}{nazwa (ścieżka) do pliku CSV z danymi}

\item{codebook}{nazwa (ścieżka) do pliku CSV z codebookiem}

\item{encoding}{ciąg znaków opisujący stronę kodową, w której zapisane są
plik z danymi i codebook (typowo "windows-1250" lub "UTF-8")}

\item{etykietyBrakowDanych}{wektor tekstowy zawierający wartości etykiet,
które mają zostać uznane za \emph{braki danych użytkownika}}
}
\value{
data.frame z etykietami nazw zmiennych i (o ile zostały zdefiniowane
w codebooku) wartości zmiennych
}
\description{
Funkcja tworzy ramkę danych ze zmiennymi \emph{etykietowanymi}
(\emph{labelled}, dokładnie klasy \emph{labelled_spss}) na podstawie pliku
CSV z danymi i pliku CSV z codebookiem (w formacie \emph{IBE'owym}).
}
\examples{
\dontrun{
dane = zastosuj_codebook("plik z danymi.csv", "codebook.csv")
naBrakiDanych = c("nie wiem", "trudno powiedzieć", "brak odpowiedzi",
  "odmowa odpowiedzi", "nie potrafię ocenić", "nie pamiętam",
  "nie wiem/trudno powiedzieć", "nie wiem/nie pamiętam")
dane = zastosuj_codebook("plik z danymi.csv", "codebook.csv", "UTF-8", naBrakiDanych)
}
}
