% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etykiety.R
\name{labels}
\alias{labels}
\alias{labels.data.frame}
\alias{labels.list}
\alias{labels.haven_labelled}
\alias{label}
\alias{label.default}
\alias{value_labels}
\alias{value_labels.data.frame}
\alias{value_labels.default}
\alias{[.val_labels_list}
\alias{as.data.frame.var_labels}
\alias{as.data.frame.val_labels}
\alias{print.var_labels}
\alias{print.val_labels}
\alias{print.val_labels_list}
\title{Operacje na etykietach zmiennych i etykietach wartosci}
\usage{
\method{labels}{data.frame}(object, SIMPLIFY = TRUE, ...)

\method{labels}{list}(object, SIMPLIFY = TRUE, ...)

\method{labels}{haven_labelled}(object, ...)

label(object, ...)

\method{label}{default}(object, ...)

value_labels(object, ...)

\method{value_labels}{data.frame}(object, ...)

\method{value_labels}{default}(object, ...)

\method{[}{val_labels_list}(x, ...)

\method{as.data.frame}{var_labels}(x, ...)

\method{as.data.frame}{val_labels}(x, ...)

\method{print}{var_labels}(x, ...)

\method{print}{val_labels}(x, ...)

\method{print}{val_labels_list}(x, ...)
}
\arguments{
\item{object}{obiekt potencjalnie posiadający zdefiniowane etykiety zmiennych
lub etykiety wartości}

\item{SIMPLIFY}{wartość logiczna - czy etykiety zmiennych mają być zwrócone
jako wektor? (jeśli \code{FALSE}, zostaną zwrócone jako lista)}

\item{...}{wyłącznie dla zachowania zgodności z funkcją \emph{generic}}

\item{x}{obiekt klasy \emph{var_labels}, \emph{var_labels} lub \emph{val_labels_list}}
}
\value{
Funkcja \code{labels} (wywołana na obiektach udokumentowanych tu
klas) zwraca listę lub wektor (z przypisaną klasą \emph{labels})
}
\description{
Funkcja \code{labels} pozwala w wygodny sposób uzyskać dostęp do etykiet
zmienych lub etykiet wartości pojedynczej zmiennej.
}
