% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etykiety.R, R/tab2.R, R/tab_n.R
\name{labels}
\alias{labels}
\alias{labels.data.frame}
\alias{labels.list}
\alias{labels.haven_labelled}
\alias{label}
\alias{label.default}
\alias{label<-}
\alias{label<-.default}
\alias{value_labels}
\alias{value_labels.data.frame}
\alias{value_labels.default}
\alias{[.val_labels_list}
\alias{as.data.frame.var_labels}
\alias{as.data.frame.val_labels}
\alias{print.var_labels}
\alias{print.val_labels}
\alias{print.val_labels_list}
\alias{label.tab_lbl2}
\alias{labels.tab_lbl2}
\alias{label.tab_lbl_n}
\alias{labels.tab_lbl_n}
\title{Operacje na etykietach zmiennych i etykietach wartosci}
\usage{
\method{labels}{data.frame}(object, SIMPLIFY = TRUE, ...)

\method{labels}{list}(object, SIMPLIFY = TRUE, ...)

\method{labels}{haven_labelled}(object, ...)

label(object, ...)

\method{label}{default}(object, ...)

label(object) <- value

\method{label}{default}(object) <- value

value_labels(object, ...)

\method{value_labels}{data.frame}(object, ...)

\method{value_labels}{default}(object, ...)

\method{[}{val_labels_list}(x, ...)

\method{as.data.frame}{var_labels}(x, ...)

\method{as.data.frame}{val_labels}(x, ...)

\method{print}{var_labels}(x, ...)

\method{print}{val_labels}(x, ...)

\method{print}{val_labels_list}(x, ...)

\method{label}{tab_lbl2}(object, ...)

\method{labels}{tab_lbl2}(object, ...)

\method{label}{tab_lbl_n}(object, ...)

\method{labels}{tab_lbl_n}(object, ...)
}
\arguments{
\item{object}{obiekt potencjalnie posiadający zdefiniowane etykiety zmiennych
lub etykiety wartości}

\item{SIMPLIFY}{wartość logiczna - czy etykiety zmiennych mają być zwrócone
jako wektor? (jeśli \code{FALSE}, zostaną zwrócone jako lista)}

\item{...}{wyłącznie dla zachowania zgodności ze wzorcem (\emph{generic})
metody}

\item{value}{wartość, która ma zostać przypisana atrybutowi \code{label}
obiektu (zwykle ciąg znaków, ale może to być dowolny obiekt)}

\item{x}{obiekt klasy \emph{var_labels}, \emph{var_labels} lub \emph{val_labels_list}}
}
\value{
Funkcja \code{labels} (wywołana na obiektach udokumentowanych tu
klas) zwraca listę lub wektor (z przypisaną klasą \emph{labels})
}
\description{
Funkcje pozwalające w wygodny sposób uzyskać dostęp do etykiet zmienych lub
etykiet wartości zarówno pojedynczej zmiennej, jak i zestawu zmiennych
w \emph{ramce danych}:
\itemize{
  \item{\code{labels} - wywołana na \emph{ramce danych} zwraca etykiety jej
        zmiennych; wywołana na obiekcie (kolumnie) klasy
        \code{haven_labelled} zwraca jego etykiety wartości,}
  \item{\code{label} - wywołana na obiekcie zwraca wartość jego atrybutu
        \code{label} - w szczególności wywołana na pojedynczej kolumnie
        \emph{ramki danych} zwraca jej etykietę zmiennej (jeśli istnieje),}
  \item{\code{label(ramka_danych$kolumna)<-"etykieta"} - pozawala przypisać
        wartość atrybutowi \code{label} obiektu - w szczególności wywołana na
        pojedynczej kolumnie \emph{ramki danych} przypisać jej etykietę
        zmiennej,}
  \item{i \code{value_labels} - wywołana na \emph{ramce danych} zwraca
        zestawienie etykiet wartości jej zmiennych;wywołana na obiekcie
        (kolumnie) klasy \code{haven_labelled} zwraca jego etykiety
        wartości.}
}
}
