% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R, R/tab_metody.R
\name{tab}
\alias{tab}
\alias{tab.data.frame}
\alias{tab.table}
\alias{tab.ftable}
\alias{tab.tbl_svy}
\alias{tab.survey.design2}
\alias{tab.svyrep.design}
\alias{tab.twophase2}
\alias{tab.default}
\alias{print.tab_lbl}
\title{Tabela z rozkladem dla zmiennych etykietowanych}
\usage{
tab(x, ..., procenty = TRUE, suma = TRUE, etykietaSuma, etykietaBD)

\method{tab}{data.frame}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA,
  w = NULL
)

\method{tab}{table}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{ftable}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{tbl_svy}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{survey.design2}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{svyrep.design}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{twophase2}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA
)

\method{tab}{default}(
  x,
  ...,
  procenty = TRUE,
  suma = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA,
  w = NULL
)

\method{print}{tab_lbl}(x, dProcenty = 1, dLiczby = 0, decimal.mark = ",", scipen = 100, ...)
}
\arguments{
\item{x}{obiekt klasy \code{labelled}, wektor, ramka danych lub obiekt
klasy \code{tbl_svy}}

\item{...}{tylko jeśli \code{x} jest \emph{ramką danych} - kolumna, której
rozkład ma zostać wygenerowany}

\item{procenty}{wartość logiczna - czy wyświetlić również rozkład częstości?}

\item{suma}{wartość logiczna - czy wyświetlić wiersz z sumą?}

\item{etykietaSuma}{ciąg znaków - etykieta dla wiersza z sumą (argument jest
ignorowany, jeśli \code{suma} równe \code{FALSE})}

\item{etykietaBD}{ciąg znaków - etykieta, którą w przygotowanym zestawieniu
mają być opisane braki danych (\code{NA}); domyślna wartość oznacza, że
zostaną one opisane jako "NA"; podanie \code{NULL} będzie skutkować
usunięciem kolumn i wierszy opisujących braki danych ze zwracanego
zestawienia}

\item{w}{opcjonalnie wektor liczbowy lub kolumna obiektu \code{x}, której
wartości zawierają wagi obserwacji, które powinny zostać uwzględnione przy
obliczaniu rozkładu}

\item{dProcenty}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami częstości}

\item{dLiczby}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami liczebności (to mogą być liczby niecałkowite,
jeśli przy tworzeniu rozkładu stosowano ważenie)}

\item{decimal.mark}{znak miejsca dziesiętnego - przekazywany do
\code{\link[base]{format}}}

\item{scipen}{liczba całkowita decydująca o skłonności R do zapisywania liczb
w notacji naukowej (czym większa, tym rzadziej R sosuje notację naukową - p.
\code{\link[base]{options}})}
}
\value{
\code{data.frame} (klasy \code{tab_lbl}) z rozkładami
}
\description{
Funkcja generuje rozkład jednej zmiennej - liczebności i częstości -
uwzględniając etykiety wartości.
}
