% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab.data.frame}
\alias{tab.default}
\alias{print.table_labeled}
\title{Tabela z rozkladem dla zmiennych etykietowanych}
\usage{
tab(x, ..., procenty = TRUE, d = 1, suma = TRUE, etykietaSuma = "ŁĄCZNIE")

\method{tab}{data.frame}(x, ..., procenty = TRUE, d = 1, suma = TRUE, etykietaSuma = "ŁĄCZNIE")

\method{tab}{default}(x, ..., procenty = TRUE, d = 1, suma = TRUE, etykietaSuma = "ŁĄCZNIE")

\method{print}{table_labeled}(x, ...)
}
\arguments{
\item{x}{obiekt klasy \code{labelled}, wektor lub ramka danych}

\item{...}{tylko jeśli \code{x} jest ramką danych - kolumna, której rozkład
ma zostać wygenerowany}

\item{procenty}{wartość logiczna - czy wyświetlić również rozkład częstości?}

\item{d}{liczba całkowita - liczba miejsc dziesiętnych, z jaką raportowane
będą procenty (argument jest ignowowany, jeśli \code{procenty} równe
\code{FALSE})}

\item{suma}{wartość logiczna - czy wyświetlić wiersz z sumą?}

\item{etykietaSuma}{ciąg znaków - etykieta dla wiersza z sumą (argument jest
ignorowany, jeśli \code{suma} równe \code{FALSE})}
}
\value{
data.frame z rozkładami
}
\description{
Funkcja generuje rozkład jednej zmiennej - liczebności i częstości -
uwzględniając etykiety wartości.
}
