% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.labelled.R
\name{tab}
\alias{tab}
\title{Tabela z rozkładem dla zmiennych etykietowanych}
\usage{
tab(x, procenty = TRUE, d = 1, suma = TRUE, etykietaSuma = "ŁĄCZNIE")
}
\arguments{
\item{x}{obiekt klasy \code{labelled}}

\item{procenty}{wartość logiczna - czy wyświetlić również rozkład częstości?}

\item{d}{liczba całkowita - liczba miejsc dziesiętnych, z jaką raportowane
będą procenty (argument jest ignowowany, jeśli \code{procenty} równe
\code{FALSE})}

\item{suma}{wartość logiczna - czy wyświetlić wiersz z sumą?}

\item{etykietaSuma}{ciąg znaków - etykieta dla wiersza z sumą (argument jest
ignorowany, jeśli \code{suma} równe \code{FALSE})}
}
\value{
data.frame z rozkładami
}
\description{
Funkcja generuje rozkład jednej zmiennej - liczebności i częstości -
uwzględniając etykiety wartości.
}
