% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2.R
\name{tab2}
\alias{tab2}
\alias{tab2.data.frame}
\alias{tab2.tbl_svy}
\alias{print.tab_lbl2}
\alias{as.data.frame.tab_lbl2}
\title{Tabela z rodzina rozkladow warunkowych dwoch zmiennych (etykietowanych)}
\usage{
tab2(x, zmW, zmK, sumowanie, liczby, procenty, etykietaSuma, etykietaBD, ...)

\method{tab2}{data.frame}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA,
  ...,
  w = NULL
)

\method{tab2}{tbl_svy}(
  x,
  zmW,
  zmK,
  sumowanie = c("brak", "kolumny", "wiersze", "ogółem"),
  liczby = TRUE,
  procenty = TRUE,
  etykietaSuma = "ŁĄCZNIE",
  etykietaBD = NA,
  ...
)

\method{print}{tab_lbl2}(x, dProcenty = 1, dLiczby = 0, decimal.mark = ",", ...)

\method{as.data.frame}{tab_lbl2}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  usunSuma = FALSE,
  wartoscBD = NULL,
  niePrzeksztalcaj = FALSE
)
}
\arguments{
\item{x}{ramka danych lub obiekt klasy \code{tbl_svy}}

\item{zmW}{kolumna obiektu \code{x}, której wartości mają zostać umieszczone
w wierszach zwróconego rozkładu}

\item{zmK}{kolumna obiektu \code{x}, której wartości mają zostać umieszczone
w kolumnach zwróconego rozkładu}

\item{sumowanie}{w którą stronę, jeśli w ogóle, powinno zostać dokonane
sumowanie lub procentowanie? (wystarczy podać pierwszą literę)}

\item{liczby}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
liczebności?}

\item{procenty}{wartość logiczna - czy zwracana tabela ma zawierać rozkłady
częstości?}

\item{etykietaSuma}{ciąg znaków - etykieta, którą w przygotowanym zestawieniu
mają być opisane braki danych \code{NA}; domyślna wartość oznacza, że zostaną
one opisane jako "NA"; podanie \code{NULL} będzie skutkować usunięciem kolumn
i wierszy opisujących braki danych ze zwracanego zestawienia}

\item{etykietaBD}{ciąg znaków - etykieta dla wiersza lub kolumy z sumą
(argument jest ignorowany, jeśli \code{sumowanie} równe \code{"brak"} lub
\code{NULL})}

\item{...}{w przypadku metody \code{print}: ew. inne parametry, przekazywane
do \code{\link[base]{format}} (wywołwanej na komórkach opisujących
liczebności lub częstości)}

\item{w}{opcjonalnie kolumna obiektu \code{x}, której wartości zawierają
wagi obserwacji, które powinny zostać uwzględnione przy obliczaniu rozkładu}

\item{dProcenty}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami częstości}

\item{dLiczby}{liczba miejsc dziesiętnych, do jakiej zostaną zaokrąglone
wartości kolumn z rozkładami liczebności (to mogą być liczby niecałkowite,
jeśli przy tworzeniu rozkładu stosowano ważenie)}

\item{decimal.mark}{znak miejsca dziesiętnego - przekazywany do
\code{\link[base]{format}}}

\item{row.names}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as.data.frame}}

\item{optional}{wyłącznie dla zgodności ze wzorcem (\emph{generic}) metody
\code{as.data.frame}}

\item{usunSuma}{wartość logiczna - czy usunąć ze zwróconej ramki danych
wiersze opisujące sumy w podgrupach?}

\item{wartoscBD}{wektor jedno- lub dwuelementowy: wartości, przy pomocy
których mają być reprezentowane ew. braki danych w zróconej ramce danych;
jeśli zostanie podany tylko jeden element, zostanie użyty zarówno
w odniesieniu do zmiennej prezentowanej w wierszach, jak i zmiennej
prezentowanej w kolumnach; jeśli zostaną podane dwa elementy, pierwszy
zostanie zastosowany do zmiennej prezentowanej w wierszach, a drugi do
zmiennej prezentowanej w kolumnach}

\item{niePrzeksztalcaj}{wartość logiczna - czy funkcja ma tylko usunąć
z obiektu klasę \code{tab_lbl2} i zwrócić go jako \emph{zwykłą} ramkę danych,
\strong{bez} dokonywania przekształcenia z postaci \emph{szerokiej} do
\emph{długiej})? (domyślnie funkcja dokonuje przekształcenia)}
}
\value{
\strong{tab2}
\code{data.frame} (klasy \code{tab_lbl2}) z rozkładami:
\itemize{
  \item{pierwsza kolumna zawiera wartości zmiennej podanej argumentem
        \code{zmW} i ma taką samą nazwę, jak ta zmienna,}
  \item{nazwy kolejnych kolumn zaczynające się od "n_" opisują rozkład
        liczebności,}
  \item{nazwy kolejnych kolumn zaczynające się od "pct_" opisują rozkład
        częstości,}
  \item{nazwy tych kolumn powstały przez połącznie przedrostka "n_" lub
        "pct_" z wartościami zmiennej podanej argumentem \code{zmK} oraz
        ew. etykietą podaną argumentem \code{etkietaSuma}.}
}
Pierwsza zwracana kolumna rozkładu co do zasady jest \emph{czynnikiem}, chyba
że jednocześnie:
\itemize{
  \item{jako argument \code{sumowanie} zostało podane "brak" lub "wiersze"
        (ew. \code{NULL}),}
  \item{nie został podany argument \code{etykietaSumy} (ew. został podany
        jako \code{NA}),}
  \item{kolumna podana argumentem \code{zmW} jest klasy "character",
        "integer", "numeric" lub "logical".}
}
W takim przypadku będzie mieć ona takiego typu, jaki kolumna miała w danych
wejściowych.
\strong{metoda as.data.frame}
W zależności od wartości parametru \code{niePrzeksztalcaj}:
\itemize{
  \item{\emph{tibble} z rozkładami przekształconymi do postaci \emph{długiej},}
  \item{obiekt klasy \code{data.frame} (i tylko tej jednej) z rozkładami
        w postacie \emph{szerokiej}.}
}
}
\description{
Funkcja generuje rozkład łączny liczebności i rozkład łączny lub rodzinę
warunkowych rozkładów częstości dwóch zmiennych. Jako pierwszy argument
przyjmuje obiekty \code{data.frame} lub \code{tbl_svy} (obiekcie będącym
połączeniem \emph{ramki danych} z informacjami o złożonym schemacie doboru
próby, tworzonym przez funkcje pakietu \emph{srvyr}). Radzi też sobie ze
zmiennymi etykietowanymi (konwertując je na \emph{czynniki}).
Metoda \code{as.data.frame} pozwala przekształcić zwracane zestawienie
w ramkę danych (ściśle w \emph{tibble}) w postaci \emph{długiej}, przydatną
np. do rysowania wykresów przy pomocy \emph{ggplot2} (czyli działa
analogicznie,jak metoda \code{as.data.frame} dla obiektów zwracanych przez
funkcję \code{\link[base]{table}} z pakietu \emph{base}).
}
